VAULT/requirements.txt
accelerate==1.10.1
beautifulsoup4==4.13.5
beir==2.1.0
comet_ml==3.47.0
entmax==1.3
evaluate==0.4.3
faiss_cpu==1.8.0.post1
filelock==3.19.1
huggingface_hub==0.34.4
ir_datasets==0.5.11
kagglehub==0.3.13
lxml==6.0.0
matplotlib==3.10.5
nltk==3.9.1
numpy==2.3.2
pandas==2.3.2
peft==0.14.0
pypdfium2==4.30.0
pyserini==0.44.0
python-dotenv==1.1.1
pytrec_eval==0.5
pytrec_eval_terrier==0.5.6
Requests==2.32.5
seaborn==0.13.2
sentence_transformers==3.0.1
tokenizers==0.21.4
torch==2.6.0
tqdm==4.67.1
transformers==4.55.4
wget==3.2
wikipedia_api==0.8.1
